Update README.md
Welcome to my collection of Python projects! This repository showcases a variety of engaging applications that demonstrate my programming skills and creativity. Each project serves a unique purpose and is a great way to explore different aspects of Python.

Projects Overview

Accident detection system is a project which detects the accident and send's a message to the emergency services with the exact location of the spot.